refactor(events): move event functions and convert process to function based options
